+++ cli2gui/argparse2json.py
@@ -2,35 +2,29 @@
 """
 import argparse
 import json
-import os
-import sys
+from os import path
+from sys import argv
 from argparse import (
 	_CountAction,
 	_HelpAction,
 	_StoreFalseAction,
 	_StoreTrueAction,
-	_SubParsersAction)
-from collections import OrderedDict
-from uuid import uuid4
+	_SubParsersAction,
+	_MutuallyExclusiveGroup)
 
-
-def is_subparser(action):
-	'''Is the element a subparser? '''
-	return isinstance(action, _SubParsersAction)
-
-def get_subparser(actions):
-	'''Get the subparser '''
-	return list(filter(is_subparser, actions))[0]
 
 def iter_parsers(parser):
 	''' Iterate over name, parser pairs '''
 	try:
-		return get_subparser(parser._actions).choices.items()
-	except:
-		return iter([('::cli2gui/default', parser)])
+		# Get the actions from the subparser
+		return [action for action in parser._actions if isinstance(
+			action, _SubParsersAction)][0].choices.items()
+	except BaseException:
+		# There is no subparser
+		return list([('::cli2gui/default', parser)])
 
 def is_default_progname(name, subparser):
-	return subparser.prog == '{} {}'.format(os.path.split(sys.argv[0])[-1], name)
+	return subparser.prog == '{} {}'.format(path.split(argv[0])[-1], name)
 
 
 def choose_name(name, subparser):
@@ -69,9 +63,6 @@ def reapply_mutex_groups(mutex_groups, action_groups):
 	return [group.update({'items': swap_actions(group['items'])}) or group
 			for group in action_groups]
 
-def is_help_message(action):
-	''' Is the element a help message? '''
-	return isinstance(action, _HelpAction)
 
 def extract_groups(action_group):
 	'''
@@ -81,20 +72,15 @@ def extract_groups(action_group):
 	return {
 		'name': action_group.title,
 		'description': action_group.description,
+		# List of items that are not help messages
 		'items': [action for action in action_group._group_actions
-				  if not is_help_message(action)],
+                    if not isinstance(action, _HelpAction)],
 		'groups': [extract_groups(group)
-				   for group in action_group._action_groups],
-
+                    for group in action_group._action_groups],
 	}
 
-def is_mutex(action):
-	return isinstance(action, argparse._MutuallyExclusiveGroup)
-
 def action_to_json(action, widget):
-
 	return {
-		'id': action.option_strings[0] if action.option_strings else action.dest,
 		'type': widget,
 		'data': {
 			'display_name': action.metavar or action.dest,
@@ -106,39 +92,32 @@ def action_to_json(action, widget):
 		},
 	}
 
+
 def build_radio_group(mutex_group, widget_group, options):
-  return {
-	'id': str(uuid4()),
-	'type': 'RadioGroup',
-	'group_name': 'Choose Option',
-	'required': mutex_group.required,
-	'data': {
-	  'commands': [action.option_strings for action in mutex_group._group_actions],
-	  'widgets': list(categorize(mutex_group._group_actions, widget_group, options))
-	}
-  }
-
-def is_flag(action):
-	""" _actions which are either, store_bool, etc.. """
-	action_types = [_StoreTrueAction, _StoreFalseAction]
-	return any(list(map(lambda Action: isinstance(action, Action), action_types)))
-
-def is_counter(action):
-	""" _actions which are of type _CountAction """
-	return isinstance(action, _CountAction)
+	return {
+		'type': 'RadioGroup',
+		'group_name': 'Choose Option',
+		'required': mutex_group.required,
+		'data': {
+			'commands': [action.option_strings for action in mutex_group._group_actions],
+			'widgets': list(categorize(mutex_group._group_actions, widget_group, options))
+		}
+  	}
 
 
 def categorize(actions, widget_dict, options):
 	for action in actions:
-		if is_mutex(action):
+		# action is of type mutex group
+		if isinstance(action, _MutuallyExclusiveGroup):
 			yield build_radio_group(action, widget_dict, options)
-		elif is_flag(action):
+		# _actions which are either, store_bool, etc..
+		elif isinstance(action, (_StoreTrueAction, _StoreFalseAction)):
 			yield action_to_json(action, "Bool")
-		elif is_counter(action):
+		# _actions which are of type _CountAction
+		elif isinstance(action, _CountAction):
 			yield action_to_json(action, "Counter")
 		else:
 			yield action_to_json(action, "TextBox")
-
 
 
 def categorize2(groups, widget_dict, options):
@@ -146,18 +125,18 @@ def categorize2(groups, widget_dict, options):
 		'name': group['name'],
 		'items': list(categorize(group['items'], widget_dict, options)),
 		'groups': categorize2(group['groups'], widget_dict, options),
-		'description': group['description'],
 	} for group in groups]
+
 
 def strip_empty(groups):
 	return [group for group in groups if group['items']]
 
+
 def process(parser, widget_dict, options):
 	mutex_groups = parser._mutually_exclusive_groups
 	raw_action_groups = [extract_groups(group) for group in parser._action_groups
-						 if group._group_actions]
+                      if group._group_actions]
 	corrected_action_groups = reapply_mutex_groups(mutex_groups, raw_action_groups)
-
 	return categorize2(strip_empty(corrected_action_groups), widget_dict, options)
 
 
@@ -171,14 +150,13 @@ def convert(parser, **kwargs):
 		dict: dictionary representing parser object
 	"""
 	return {
-		'widgets': OrderedDict(
-			(choose_name(name, sub_parser), {
-				'command': name,
+		'widgets': [
+			{
 				'name': choose_name(name, sub_parser),
 				'help': get_subparser_help(sub_parser),
-				'description': '',
 				'contents': process(sub_parser,
 					getattr(sub_parser, 'widgets', {}),
 					getattr(sub_parser, 'options', {}))
-			}) for name, sub_parser in iter_parsers(parser))
+			} for name, sub_parser in iter_parsers(parser)
+		]
 	}