+++ CHANGELOG.md
@@ -2,5 +2,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.1 - 2020/03/12
+- Updated readme
+- added images
+- added getopt parser
+- Program icon is to left of title
+- Streamlined argparse2json
+
 ## 2020 - 2020/03/06
 - First release